#31 - Beginning finalisation of hangar scene and features, as well as final refactor
