+++ client/src/app/scenograph/actors/player.js
@@ -53,7 +53,7 @@ export default class Player {
     async load() {
 
         // Setup aircraft, used for the intro sequence.
-        this.vehicle = new l.scenograph.objects.vehicles.valiant();
+        this.vehicle = new l.scenograph.objects.vehicles.valiant(this.actorInstance);
         await this.vehicle.load();
         l.current_scene.scene.add(
           this.vehicle.mesh
