--- game/src/objects/aircraft/valiant.ts
@@ -1,15 +1,0 @@
-/**
- * Valiant Aircraft, default ship and Kingdom of Winthrom main vehicle
- */
-
-import BaseAircraft from './base';
-
-class Valiant extends BaseAircraft {
-
-    constructor( mesh ) {
-        super( mesh ); // Call the constructor of the base class
-    }
-
-}
-
-module.exports = Valiant;
