+++ game/src/objects/vehicles/valiant.ts
@@ -1,17 +1,16 @@
 /**
- * Person class
+ * Valiant Aircraft, default ship and Kingdom of Winthrom main vehicle
  */
 
-import ObjectBase from './base';
-import BaseActor from '../actors/base2';
-import { changeVelocity, normaliseSpeedDelta, easeOutExpo, easeInQuad, easeInOutExpo } from '../helpers';
-import { Vec3 } from '../types';
+import ObjectBase from '../base';
+import ActorBase from '../actors/base2';
 
+import { changeVelocity, normaliseSpeedDelta, easeOutExpo, easeInQuad, easeInOutExpo } from '../../helpers';
 
-export default class Person extends ObjectBase {
+class Valiant extends ObjectBase {
 
     // Actor that controls this object.
-    public actor?:          BaseActor;
+    public actor?:          ActorBase;
 
     // Object world parameters
     public hitPoints:       number                              = 100;
@@ -22,63 +21,58 @@ export default class Person extends ObjectBase {
     public maxUp:           number                              = 4 / 60;
     public maxDown:         number                              = 16 / 60;  // gravity?
 
-    constructor() {
-        super();
-        this.aabb = {
-            halfSize: { x: 0.5, y: 0.75, z: 0.3 }
-        };
+    constructor( mesh ) {
+        super( mesh ); // Call the constructor of the base class
     }
 
-    update( time_delta: number  ) {
+    /**
+     * Move the aircraft based on velocity, direction and time delta between frames.
+     *
+     * @param time_delta
+     */
+    public update( time_delta: number ): object {
         if (!this.actor) return;
-        this.nextPosition = { ...this.position }; // start with current position
-
-        let stepSize:           number = .025 * normaliseSpeedDelta( time_delta ),
+        let stepSize:           number = .05 * normaliseSpeedDelta( time_delta ),
             rY:                 number = 0,
             tZ:                 number = 0,
             tY:                 number = 0,
-            radian:             number = - (Math.PI / 180) * stepSize * 100;
+            radian:             number = (Math.PI / 180);
 
-        if ( this.actor.controls.forward || this.actor.controls.back ){
-            // Update Airspeed (horizontal velocity)
-            this.airSpeed = changeVelocity(
-                stepSize * easeInOutExpo( 1 - ( Math.abs ( this.airSpeed ) / this.maxForward ) ),
-                stepSize,
-                this.airSpeed,
-                this.actor.controls.forward,
-                this.actor.controls.back,
-                this.maxForward,
-                this.maxBackward,
-                easeOutExpo( 0.987 )
-            );
-        }
-        else {
-            this.airSpeed = 0;
-        }
+        // Update Airspeed (horizontal velocity)
+        this.airSpeed = changeVelocity(
+            stepSize * easeInOutExpo( 1 - ( Math.abs ( this.airSpeed ) / this.maxForward ) ),
+            stepSize,
+            this.airSpeed,
+            this.actor.controls.throttleUp,
+            this.actor.controls.throttleDown,
+            this.maxForward,
+            this.maxBackward,
+            easeOutExpo( 0.987 )
+        );
 
         // Update Vertical Speed (velocity)
         this.verticalSpeed = changeVelocity(
             stepSize * easeInOutExpo( 1 - ( Math.abs ( this.verticalSpeed ) / this.maxUp ) ),
             stepSize * easeInOutExpo( 1 - ( Math.abs ( this.verticalSpeed ) / this.maxDown ) ),
             this.verticalSpeed,
-            this.actor.controls.crouch,     // Note: Move Down/Up is reversed by design.
-            this.actor.controls.jump,
+            this.actor.controls.moveDown,     // Note: Move Down/Up is reversed by design.
+            this.actor.controls.moveUp,
             this.maxDown,
             this.maxUp,
             easeInQuad( 0.321 )
         );
 
-        // Check the vertical speed exceeds minimum threshold for change in vertical position
-        if (Math.abs(this.verticalSpeed) > 0.01) {
+         // Check the vertical speed exceeds minimum threshold for change in vertical position
+         if (Math.abs(this.verticalSpeed) > 0.01) {
             tY = this.verticalSpeed;
         }
 
         // Turning
-        if (this.actor.controls.turnRight) {
+        if (this.actor.controls.moveLeft) {
             rY += radian;
         }
         else {
-            if (this.actor.controls.turnLeft) {
+            if (this.actor.controls.moveRight) {
                 rY -= radian;
             }
         }
@@ -91,6 +85,14 @@ export default class Person extends ObjectBase {
 
         }
 
+        // Animate the ship's rotation in the game client based on controls.
+        if (
+            !(this.actor.controls.throttleDown || this.actor.controls.throttleUp) &&
+            !(this.actor.controls.moveDown || this.actor.controls.moveUp)
+        ) {
+            this.rotation.x *= .9;
+        }
+
         if (rY != 0) {
             if (Math.abs(this.rotation.z) < Math.PI / 4) {
                 this.rotation.z += rY / Math.PI;
@@ -98,19 +100,22 @@ export default class Person extends ObjectBase {
 
             this.rotation.y += rY;
         }
+        else {
+            this.rotation.z *= .9;
+        }
 
         let xDiff = tZ * Math.sin(this.rotation.y),
             zDiff = tZ * Math.cos(this.rotation.y);
 
         // "1" is the floor limit as it's the ocean surface and the camera clips through the water any lower.
-        if (this.nextPosition.y + tY >= 1 ) {
-            this.nextPosition.y += tY;
+        if (this.position.y + tY >= 1 ) {
+            this.position.y += tY;
         } else {
             this.verticalSpeed = 0;
         }
 
-        this.nextPosition.x += xDiff;
-        this.nextPosition.z += zDiff;
+        this.position.x += xDiff;
+        this.position.z += zDiff;
 
         this.rY = rY;
         this.tY = tY;
@@ -118,8 +123,6 @@ export default class Person extends ObjectBase {
 
     }
 
-    commitNextPosition() {
-        this.position = this.nextPosition;
-    }
-
 }
+
+module.exports = Valiant;
